fix: auto-resize
